build: Rework image tagging
We should really create proper multi-arch images, ... but we can at
least improve the current system we have of tagging the image with the
host arch.

The problem with the way we do it now is that the two tags (the plain
one and the one with the arch) cause docker rmi to refuse to delete the
image until one if explicitly untagged.

Instead we can only create the tag with the arch name in it temporarily,
when we're pulling/pushing.

To make that work with the allcross image we need to drop --pull, so
that the FROM will use the local image without the arch name.