Dis-ambiguate the container registry to pull images from
Podman requires the user to be explicit about the container registry
from which to pull an image from. In the default setup, trying to pull
the linuxppc build container throws the below error:

  $ make pull-image@ppc64le@ubuntu@21.10
  + podman pull linuxppc/build:ubuntu-21.10-x86_64
  Error: short-name "linuxppc/build:ubuntu-21.10-x86_64" did not resolve
  to an alias and no unqualified-search registries are defined in
  "/etc/containers/registries.conf"
  Error: pulling linuxppc/build:ubuntu-21.10?

While users can setup the docker hub as the default unqualified search
registry, it is easier to simply be explicit about the registry in the
first place. Prefix all images with docker.io for this purpose.