Fix detection of pulled images
While docker can pull images when prefixed with the registry url, it
doesn't include the registry names in the local image repository. Due to
this, trying to detect if an image is present in the local repository
fails causing us to try and build the image again:
	$ make SRC=~/linux kernel@ppc64le@fedora@34
	+ docker images -q --filter=reference=docker.io/linuxppc/build:fedora-34
	+ docker build --pull -f fedora/Dockerfile ... <snip>

Fix this by only appending the registry url for podman.