build: Add korg v13.1.0 image
Gcc v13.1.0 cross-toolchain is available from kernel.org. Add support
for building the kernel with it. We have to use ubuntu 23.04 due to
library dependencies from the build.

Ubuntu 23.04 also now adds a default user 'ubuntu' with uid 1000. That
uid is commonly used in user accounts on most distros, so it causes an
issue with the subsequent useradd command in the Dockerfile. Delete the
'ubuntu' user as a workaround for the same.