build: Symlink to the latest build
With numerous build options, it is not always easy to guess/remember the
output directory for the most recent build. Have the scripts add a
symlink with the name latest-[build type] (latest-kernel,
latest-selftests, latest-perf) pointing to the last successful build in
the $CI_OUTPUT directory.  This makes it very easy to refer to artifacts
from the most recent build.